requirements.txt
alembic==1.8.0
blinker==1.4
certifi==2022.6.15
chardet==3.0.4
click==8.1.3
Flask==1.1.1
Flask-API==2.0
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gitdb==4.0.9
gitdb2==4.0.2
GitPython==3.1.3
greenlet==1.1.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.1
pycrypto==2.6.1
PyJWT==1.7.1
python-dotenv==0.20.0
PyYAML==5.3.1
requests==2.24.0
simple-crypt==4.1.7
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.39
urllib3==1.25.11
uuid==1.30
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.8.0
